refactored localchecklist primitive into a store
